Fixing composer issues.
Someone (naming no names and not looking at Finlay _specifically_) had ran some composer
commands on a php 7.2 machine which bumped the versions of various composer packages to
require 7.2.

But the server is on 7.1 - so this would have blown up quite a bit if deployed.  Noticed
while changing the wording of some emails for Lynsey in the HR/finance office.